完善数据库删除的条件

+++ src/main/java/SqlFunction/DropDatabase.java
@@ -1,6 +1,7 @@
 package SqlFunction;
 
 import java.io.*;
+import java.util.Scanner;
 
 public class DropDatabase {
     public DropDatabase(String DatabaseName){
@@ -12,14 +13,28 @@ public class DropDatabase {
         }else{
             //将该文件夹下的文件全部删除
             File[] files = dir.listFiles();
-            for(int i = 0; i < files.length; i++){
-                //按顺序删除文件
-                flag = files[i].delete();
-                //如果删除失败
-                if(!flag) break;
+            if(files.length > 0){
+                System.out.println("数据库中有表格，是否继续删除？（Y/N）");
+
+                Scanner scanner = new Scanner(System.in);
+                String answer = scanner.next();
+                if(answer.toUpperCase().equals('Y')){
+                    for(int i = 0; i < files.length; i++){
+                        //按顺序删除文件
+                        flag = files[i].delete();
+                        //如果删除失败
+                        if(!flag) break;
+                    }
+                    dir.delete();
+                    System.out.println("数据库"+DatabaseName+"删除成功！");
+                }else{
+                    System.out.println("数据库"+DatabaseName+"未被删除！");
+                }
+            }else{
+                //数据库为空，直接删除数据库
+                dir.delete();
+                System.out.println("数据库"+DatabaseName+"删除成功！");
             }
-            dir.delete();
-            System.out.println("数据库"+DatabaseName+"删除成功！");
         }
     }
 }
